Add controller for friendly matches
